requirements.txt
ffmpeg_python==0.2.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
